+++ tests/bor/bor_reorg_test.go
@@ -3,161 +3,27 @@
 package bor
 
 import (
-	"crypto/ecdsa"
-	"encoding/json"
-	"io/ioutil"
-	"math/big"
-	"os"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/ethereum/go-ethereum/accounts/keystore"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/fdlimit"
-	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/ethereum/go-ethereum/eth"
-	"github.com/ethereum/go-ethereum/eth/downloader"
-	"github.com/ethereum/go-ethereum/eth/ethconfig"
-	"github.com/ethereum/go-ethereum/log"
-	"github.com/ethereum/go-ethereum/miner"
-	"github.com/ethereum/go-ethereum/node"
-	"github.com/ethereum/go-ethereum/p2p"
-	"github.com/ethereum/go-ethereum/p2p/enode"
-	"github.com/ethereum/go-ethereum/params"
 	"github.com/stretchr/testify/assert"
 )
 
-var (
-	// addr1 = 0x71562b71999873DB5b286dF957af199Ec94617F7
-	pkey1, _ = crypto.HexToECDSA("b71c71a67e1177ad4e901695e1b4b9ee17ae16c6668d313eac2f96dbcda3f291")
-	// addr2 = 0x9fB29AAc15b9A4B7F17c3385939b007540f4d791
-	pkey2, _ = crypto.HexToECDSA("9b28f36fbd67381120752d6172ecdcf10e06ab2d9a1367aac00cdcd6ac7855d3")
-	keys     = []*ecdsa.PrivateKey{pkey1, pkey2}
-)
-
-func initMiner(genesis *core.Genesis, privKey *ecdsa.PrivateKey) (*node.Node, *eth.Ethereum, error) {
-	// Define the basic configurations for the Ethereum node
-	datadir, _ := ioutil.TempDir("", "")
-
-	config := &node.Config{
-		Name:    "geth",
-		Version: params.Version,
-		DataDir: datadir,
-		P2P: p2p.Config{
-			ListenAddr:  "0.0.0.0:0",
-			NoDiscovery: true,
-			MaxPeers:    25,
-		},
-		UseLightweightKDF: true,
-	}
-	// Create the node and configure a full Ethereum node on it
-	stack, err := node.New(config)
-	if err != nil {
-		return nil, nil, err
-	}
-	ethBackend, err := eth.New(stack, &ethconfig.Config{
-		Genesis:         genesis,
-		NetworkId:       genesis.Config.ChainID.Uint64(),
-		SyncMode:        downloader.FullSync,
-		DatabaseCache:   256,
-		DatabaseHandles: 256,
-		TxPool:          core.DefaultTxPoolConfig,
-		GPO:             ethconfig.Defaults.GPO,
-		Ethash:          ethconfig.Defaults.Ethash,
-		Miner: miner.Config{
-			Etherbase: crypto.PubkeyToAddress(privKey.PublicKey),
-			GasCeil:   genesis.GasLimit * 11 / 10,
-			GasPrice:  big.NewInt(1),
-			Recommit:  time.Second,
-		},
-		WithoutHeimdall: true,
-	})
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// register backend to account manager with keystore for signing
-	keydir := stack.KeyStoreDir()
-
-	n, p := keystore.StandardScryptN, keystore.StandardScryptP
-	kStore := keystore.NewKeyStore(keydir, n, p)
-
-	kStore.ImportECDSA(privKey, "")
-	acc := kStore.Accounts()[0]
-	kStore.Unlock(acc, "")
-	// proceed to authorize the local account manager in any case
-	ethBackend.AccountManager().AddBackend(kStore)
-
-	// ethBackend.AccountManager().AddBackend()
-	err = stack.Start()
-	return stack, ethBackend, err
-}
-
-func initGenesis(t *testing.T, faucets []*ecdsa.PrivateKey) *core.Genesis {
-
-	// sprint size = 8 in genesis
-	genesisData, err := ioutil.ReadFile("./testdata/genesis_2val.json")
-	if err != nil {
-		t.Fatalf("%s", err)
-	}
-
-	genesis := &core.Genesis{}
-
-	if err := json.Unmarshal(genesisData, genesis); err != nil {
-		t.Fatalf("%s", err)
-	}
-
-	genesis.Config.ChainID = big.NewInt(15001)
-	genesis.Config.EIP150Hash = common.Hash{}
-
-	return genesis
-}
-
 func TestValidatorWentOffline(t *testing.T) {
 
-	log.Root().SetHandler(log.LvlFilterHandler(log.LvlInfo, log.StreamHandler(os.Stderr, log.TerminalFormat(true))))
-	fdlimit.Raise(2048)
-
-	// Generate a batch of accounts to seal and fund with
-	faucets := make([]*ecdsa.PrivateKey, 128)
-	for i := 0; i < len(faucets); i++ {
-		faucets[i], _ = crypto.GenerateKey()
-	}
-
 	// Create an Ethash network based off of the Ropsten config
-	genesis := initGenesis(t, faucets)
+	genesis := initGenesis(t)
+	stacks, nodes, enodes := setupMiner(t, 2, genesis)
 
-	var (
-		stacks []*node.Node
-		nodes  []*eth.Ethereum
-		enodes []*enode.Node
-	)
-	for i := 0; i < 2; i++ {
-		// Start the node and wait until it's up
-		stack, ethBackend, err := initMiner(genesis, keys[i])
-		if err != nil {
-			t.Fatal("Error occured while initialising miner", "error", err)
+	defer func() {
+		for _, stack := range stacks {
+			stack.Close()
 		}
-		defer stack.Close()
-
-		for stack.Server().NodeInfo().Ports.Listener == 0 {
-			time.Sleep(250 * time.Millisecond)
-		}
-		// Connect the node to all the previous ones
-		for _, n := range enodes {
-			stack.Server().AddPeer(n)
-		}
-		// Start tracking the node and its enode
-		stacks = append(stacks, stack)
-		nodes = append(nodes, ethBackend)
-		enodes = append(enodes, stack.Server().Self())
-	}
+	}()
 
 	// Iterate over all the nodes and start mining
-	time.Sleep(3 * time.Second)
 	for _, node := range nodes {
 		if err := node.StartMining(1); err != nil {
 			t.Fatal("Error occured while starting miner", "node", node, "error", err)
@@ -320,17 +186,9 @@ func TestForkWithBlockTime(t *testing.T) {
 			forkExpected:  true,
 		},
 	}
-	log.Root().SetHandler(log.LvlFilterHandler(log.LvlInfo, log.StreamHandler(os.Stderr, log.TerminalFormat(true))))
-	fdlimit.Raise(2048)
-
-	// Generate a batch of accounts to seal and fund with
-	faucets := make([]*ecdsa.PrivateKey, 128)
-	for i := 0; i < len(faucets); i++ {
-		faucets[i], _ = crypto.GenerateKey()
-	}
 
 	// Create an Ethash network based off of the Ropsten config
-	genesis := initGenesis(t, faucets)
+	genesis := initGenesis(t)
 
 	for _, test := range cases {
 		genesis.Config.Bor.Sprint = test.sprint
@@ -338,35 +196,15 @@ func TestForkWithBlockTime(t *testing.T) {
 		genesis.Config.Bor.BackupMultiplier = test.blockTime
 		genesis.Config.Bor.ProducerDelay = test.producerDelay
 
-		var (
-			stacks []*node.Node
-			nodes  []*eth.Ethereum
-			enodes []*enode.Node
-		)
+		stacks, nodes, _ := setupMiner(t, 2, genesis)
 
-		for i := 0; i < 2; i++ {
-			// Start the node and wait until it's up
-			stack, ethBackend, err := initMiner(genesis, keys[i])
-			if err != nil {
-				t.Fatal("Error occured while initialising miner", "error", err)
+		defer func() {
+			for _, stack := range stacks {
+				stack.Close()
 			}
-			defer stack.Close()
-
-			for stack.Server().NodeInfo().Ports.Listener == 0 {
-				time.Sleep(250 * time.Millisecond)
-			}
-			// Connect the node to all the previous ones
-			for _, n := range enodes {
-				stack.Server().AddPeer(n)
-			}
-			// Start tracking the node and its enode
-			stacks = append(stacks, stack)
-			nodes = append(nodes, ethBackend)
-			enodes = append(enodes, stack.Server().Self())
-		}
+		}()
 
 		// Iterate over all the nodes and start mining
-		time.Sleep(3 * time.Second)
 		for _, node := range nodes {
 			if err := node.StartMining(1); err != nil {
 				t.Fatal("Error occured while starting miner", "node", node, "error", err)
@@ -375,22 +213,31 @@ func TestForkWithBlockTime(t *testing.T) {
 
 		var wg sync.WaitGroup
 		blockHeaders := make([]*types.Header, 2)
+		ticker := time.NewTicker(time.Duration(test.blockTime["0"]) * time.Second)
 
 		for i := 0; i < 2; i++ {
 			wg.Add(1)
 
 			go func(i int) {
 				defer wg.Done()
+
 				for {
-					blockHeaders[i] = nodes[i].BlockChain().GetHeaderByNumber(test.sprint*test.change + 10)
-					if blockHeaders[i] != nil {
-						break
+					select {
+					case <-ticker.C:
+						blockHeaders[i] = nodes[i].BlockChain().GetHeaderByNumber(test.sprint*test.change + 10)
+						if blockHeaders[i] != nil {
+							return
+						}
+					default:
+
 					}
 				}
+
 			}(i)
 		}
 
 		wg.Wait()
+		ticker.Stop()
 
 		// Before the end of sprint
 		blockHeaderVal0 := nodes[0].BlockChain().GetHeaderByNumber(test.sprint - 1)
